+++ client/js/components/statement/publicStatementModal/StatementModalRecheck.vue
@@ -262,7 +262,7 @@
       :class="prefixClass('flow-root border--top u-pt-0_25')"
     >
       <span :class="prefixClass('flow-root')">
-        <em>{{ Translator.trans('more.data') }}: </em>
+        <em>{{ Translator.trans('statement.data') }}: </em>
         <button
           type="button"
           data-cy="statementModalRecheck:customFieldsEdit"
@@ -278,37 +278,33 @@
         </button>
       </span>
 
-      <dl>
-        <div
-          v-for="field in customFieldsWithValues"
-          :key="field.id"
-          :class="prefixClass('mb-1')"
-        >
-          <dt :class="prefixClass('weight--bold')">
-            {{ field.name }}
-          </dt>
-          <dd :class="prefixClass('ml-2')">
-            <span
-              v-for="option in field.selected"
-              :key="option.id"
-              :class="prefixClass('block')"
-            >
-              {{ option.label }}
-            </span>
-          </dd>
-        </div>
-      </dl>
+      <custom-field
+        v-for="field in customFieldsWithValues"
+        :key="field.id"
+        :definition-source-id="procedureId"
+        :field-data="{ id: field.id, value: field.value || [] }"
+        mode="readonly"
+        :class="prefixClass('mb-1')"
+      >
+        <template v-slot:readonly-display="{ field: customField }">
+          <span>
+            {{ customField.value.selectedOptions.map(o => o.label).join(', ') }}
+          </span>
+        </template>
+      </custom-field>
     </div>
   </fieldset>
 </template>
 
 <script>
 import { CleanHtml, DpInlineNotification, prefixClassMixin } from '@demos-europe/demosplan-ui'
+import CustomField from '@DpJs/components/customFields/CustomField'
 
 export default {
   name: 'StatementModalRecheck',
 
   components: {
+    CustomField,
     DpInlineNotification,
   },
 
@@ -371,6 +367,12 @@ export default {
       required: false,
       default: () => [],
     },
+
+    procedureId: {
+      type: String,
+      required: false,
+      default: '',
+    },
   },
 
   emits: [
@@ -408,7 +410,7 @@ export default {
 
     customFieldsWithValues () {
       return this.selectableCustomFields.filter(
-        field => field.selected && field.selected.length > 0,
+        field => field.value && field.value.length > 0,
       )
     },
   },